Update patch set 1
Patch Set 1:

I would categorize it as "data generation" script, like the docker-images-collector.sh Its outcome will be a list that will serve as input for downloading. What is missing is to know what pom.xml we need (as far as I know we do not have automated way to identify them).

Patch-set: 1