Always show healthcheck output if it was run
This patch changes the default behavior of showing
healthcheck script results. So far output was printed
only in increased verbosity mode but the proper action
should be to print it whenever it was called.

A bug in exec_healthcheck() routine is also fixed
(subprocess.check_output() returns byte string, not
an object with 'output' attribute).

Change-Id: Ifd7294de9bed321123f74977f881999ce69cf878
Issue-ID: OOM-1806
Signed-off-by: Bartek Grzybowski <b.grzybowski@partner.samsung.com>