Add kubernetes-node group for workers
Moved all kubernetes nodes from the kubernetes to kubernetes-node group,
and made the kubernetes group a parent of both kubernetes-node and
kubernetes-control-plane.
The reason is that we still need to have separate groups for
control planes and for nodes, but some
operations are performed equally on any kind of kubernetes
cluster member, and currently one would need to separately include all
related groups one by one. Example of
such common operation is installation of docker, that was not
performed for kubernetes-control-plane group, so that
docker was not installed on control planes that were neither infra nor nodes.
The side effect of this change is that if infrastructure
server is also a cluster member, it lands in a kubernetes group by default,
and node specific setup would be performed on it.
For that reason, playbook is modified to perform this setup on all
kubernetes cluster members except the infra server.

Change-Id: Ic827002d28e535334dbab7e5ad4aed8aa95c97a9
Issue-ID: OOM-1778
Signed-off-by: Michal Zegan <m.zegan@samsung.com>