Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > Actually wondering if it doesn't also make sense to add the kubernetes-etcd group? in case of production clusters you rather don't colocate etcd and nodes, and it is now relatively easy to do.
> 
> Or is the kubernetes-control-plane group actually a kubernetes-master group now that would be equal to kubernetes-etcd? So can deal both with just kubernetes-master group?
Well they are separate roles not without a reason. Especially if we want to have high availability support in the future, we need a way to have etcd cluster separate from both master and nodes (especially nodes).
It may be something not so relevant here, but it seems a natural extension.

Patch-set: 1