Update patch set 1
Patch Set 1:

This commit can be regarded as a final in the stack of related ones. Each of the commits in the pile is atomic and introduces some isolated change but i think it's sufficient to verify this one only as it's a final, working version. Change in functionality (at least that visible to user) is described in 88855.

Patch-set: 1
Reviewer: Gerrit User 5097 <5097@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0