Do not use bindmounts for nfs tests
This commit removes bindmounting
of host directories to containers in play-resources tests.
It seems that it is unneeded, and it is enough to just create anonymous volumes
instead of cluttering host filesystem.
This is one part of a preparation for removing the cleanup-directories
role that won't be needed anymore.
Note: similar changes were made to the nfs
role tests, for the same reasons, but
cleanup-directories is not used there

Change-Id: Ib33543427d72b4bc7ede532e5eb9a5023931f1fe
Issue-ID: OOM-1902
Signed-off-by: Michal Zegan <m.zegan@samsung.com>