Simplify the helm override file handling
This is a followup for change 89022.
It simplifies things by not including the app helm override in the overrides
list directly. Instead it is always appended last
by application-override role after it's generated at runtime.
Also the variable with overrides has a new name and is an empty list by
default.

Change-Id: Id7f448922905ecd6d4434db87058dd17b2f34d9c
Issue-ID: OOM-1861
Signed-off-by: Michal Zegan <m.zegan@samsung.com>