Modify the way override works
Now the generated app_helm_override_file and the list of
helm overrides is separate.
The former is to contain things like user provided overrides,
the latter is the list of paths to preexisting helm value files, by default
relative to helm chart dirs, but absolute
paths are supported too if needed.
The playbook joins all that into a final list of override files that is then
passed to helm.

Change-Id: I8b8ded38b39aa9a75e55fc63fa0e11b986556cb8
Issue-ID: OOM-1951
Signed-off-by: Michal Zegan <m.zegan@samsung.com>