Add helm redeploy tool
Simple script to try to redeploy failed charts or explicitly desired
chart(s).

It supports override files for helm - but please, beware that helm
is not capable of handling whitespace in filenames so I don't support
that in the script too (I made a version which does but it is no use if
helm will break anyway).

Change-Id: I69bf0672a64c45e30d4e4239da561647d00637a7
Issue-ID: OOM-1984
Signed-off-by: Petr Ospalý <p.ospaly@partner.samsung.com>