Update patch set 1
Patch Set 1:

I tested it, it does what it claims (for me at least) - no silver bullet. There are maybe bugs (please try to use it and let me know). I wanted to improve deployment and made it more granular and parallel with helm install per chart but I am not confident if it will be the same as with the whole helm deploy onap - until I have a fully functional env I cannot verify anything.

Also I made an update to this script which supported whitespace in override filenames (it was ugly hack because shell...) and it worked but unfortunately the helm itself cannot handle whitespace, so I dropped the change and left the script more readable.

Patch-set: 1