Add onap instance configuration role
This change adds a role that configures
deployed onap infrastructure instances. This role is generic
and not openstack dependent.
This role configures instances to reach each other by name, to mutually
authenticate with the given ssh keys, etc. This
mainly makes it easier for developers to do troubleshooting, and
allows offline installer to work smoothly.

Change-Id: I26750b0a8a69f56297f0a5f7ff491463ed2e9e32
Issue-ID: OOM-2042
Signed-off-by: Michal Zegan <m.zegan@samsung.com>