Add inventory for cicdansible playbook
This change adds the basic inventory file and group_vars used when running
cicdansible playbook to deploy onap on openstack.
This inventory contains only instances that must be present, other inventory
entries are dynamically registered during install.
Also, it delegates to group_vars for things like floating ip
addresses.
group_vars contains default settings for the playbook, and especially the
all.yml must be modified for the environment to be deployed.

Change-Id: Ia93e6b98841d7edc45699b44b2479af6daff6c4f
Issue-ID: OOM-2042
Signed-off-by: Michal Zegan <m.zegan@samsung.com>