Add ability to put dockerdata-nfs on root disk
Modify the heat template so that it becomes possible to
skip volume creation for dockerdata-nfs.
This feature can be toggled at will. It is useful for the cases
where /dockerdata-nfs needs to be fast, but there are no ssd volumes available.
Of course, large enough ephemeral disks are needed, around
60 gb.

Change-Id: I37e29e7cf7273d1d4a6e896ec017d9ef8177b1d0
Issue-ID: OOM-2042
Signed-off-by: Michal Zegan <m.zegan@samsung.com>