Update patch set 7
Patch Set 7:

> Patch Set 7: Code-Review-2
> 
> (1 comment)
> 
> A general question: what's the point of upstreaming our internal CI pipeline config? Shouldn't it stay in our internal repo? In it's current shape it's useless for anyone but us.

The plan is to enable CI in upstream ONAP LF servers and promote it to a official ONAP nighly build. Same way as easy-installer would be an official (and only actually feasible) way to install onap either offline or online way.

This requires that integration team will use that nighly in official testing. 

If there's anything else in mind, then I agree that this can be removed.

Patch-set: 7