Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1: Code-Review-1
> > 
> > Testing onap demos is content that does not directly belong to offline-installer and even less into cicdansible tool, where this commit is placing it.
> > 
> > Would be better to host these test in onap's demo repository. Maybe implement it as Ansible-galaxy collection https://galaxy.ansible.com/docs/contributing/creating_collections.html. Anyway separated from both cicdansible and offline-installer ansible code base.
> 
> If I understand my task correctly, I am not doing vFWCL itself, but only jenkins pipeline for it and only in our jenkins. Then Onap's demo repo is not good place. However, ci-management repo neither as this pipeline will not be present in onap upstream jenkins.
> 
> If you persist on putting it ci-management repo, shouldn't be then whole cicdansible tool also placed in onap's ci-management repository?

Yes it's not vFW itself but support to run automatically vFW demo in the context of offline onap deployment (or prob/hopefully also online onap).

Our pipeline should be present in onap upstream Jenkins (that has been all the time our target).

cicdansible tool is already upstreamed and better place for it could be in ci-management. Expecially if it's somehow impossible otherwise to get pipeline to onap Jenkins. But I don't in general see reason why cicdansible could not be also in current repo.

But the point of my comment, that where this running onap demo code should be, is still same. It does not belong to cicdansible, offline-installer code base directly, but is content that can be executed by these tools. And content very tightly is related to the demo itself.


One more tactical comment. The code seems to be very simple and short so there's prob no need to create roles out of this as there's no usage of those roles elsewhere. So whole demo run maybe could be a single ansible playbook having all code.

Patch-set: 1