Update patch set 5
Patch Set 5:

Samuli is adding "changed_when: false" to this task to fake it's idempotence is ok with you?

For this particular issue (which this commit is addressing) we're not able to provide a clean solution that wouldn't be more or less hacky so there needs to be some kind of compromise in whatever way we deal with it.

Patch-set: 5