Update patch set 5
Patch Set 5:

How idempotent was that archive module? Do we know? We could make this idempotent too, by first checking in the destination that these helm_charts are there (if not the continue), then create archive there and store checksum of it in some ansible variable and throw that archive away, then do the actual tar in the source, make checksum of it and compare two checksums, and act on the result.

So it is possible in more than one way to satisfy this idempotency, but do we care? Let's say: it is idempotent and on the destination is outdated version of helm_charts - does ansible archive module rewrite it with newer version? I think yes, so the only benefit of idempotency is that it tells us: "changed" and not just "ok"

am I right?

Patch-set: 5
Label: Code-Review=0