Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> How idempotent was that archive module? Do we know?

Yes, we do: archive module deals with it internally by creating working archive and comparing it's uniqueness property with resulting archive (if it already exists) and returns "not changed" in case they're the same

> We could make this idempotent too, by first checking (...)

In this review we're discussing creating archive idempotency so i find your remarks as utterly valid but out of scope as you're referencing the whole process idempotency, not this task alone. Please also note that unpacking process is done by "unarchive" module which is idempotent on it's own so i don't see an issue here.

> So it is possible in more than one way to satisfy this idempotency, but do we care? Let's say: it is idempotent and on the destination is outdated version of helm_charts - does ansible archive module rewrite it with newer version? I think yes, so the only benefit of idempotency is that it tells us: "changed" and not just "ok"
> 
> am I right?

My above comment also addresses this one of yours - whatever happens on dest is unarchive's module job and out of scope of this changeset i think.

Patch-set: 5
Label: Code-Review=0