Update patch set 7
Patch Set 7:

> Patch Set 7:
> 
> @bartek I was thinking that archive can also transfer to another machine and that is what we are doing, but if it just creates a tra next to a tarred directorz, then it is even simpler and my ide of using working/temp tar is the same in both, unarchive module should in the case of the idempotencz untar the archive onlz if there are changes - so that is why I dont see it as independent steps

Unarchive module has the transferring capability and that's how we use it in "Unarchive helm charts on infra node" task.

Performing "unarchive" task based on changed status of "archive" task is not legitimate because unarchive has it's own idempotency logic and it has to be evaluated each time (i.e. it has to be always run because it decides upon it's own idempotency logic whether extract files on remote or not; please see idempotency remarks in https://docs.ansible.com/ansible/latest/modules/unarchive_module.html)

Patch-set: 7