Update patch set 7
Patch Set 7:

> Patch Set 7:
> 
> adding "changed when false" was not my suggestion. Even that can be used to fool molecule test. But the point was that molecule idempotence test is disabled for this whole module. And at some point we could fix it?
> For user it would matter more that playbook just can be run again if it fails (not real idempotency meaning it only makes changes when it's not yet done). Is this now re-runnable?

While I agree with your remarks please note that this patch is in no way targeted at fixing idempotency in the scope of whole application.yml and wasn't supposed to, it's different topic. Idempotency matter here popped out only in the scope of the changed task code (helm charts archival)

Please raise a ticket for it if you'd like to see it fixed at some point.

Patch-set: 7