Update patch set 7
Patch Set 7:

> Patch Set 7:
> 
> > Patch Set 7:
> > 
> > adding "changed when false" was not my suggestion. Even that can be used to fool molecule test. But the point was that molecule idempotence test is disabled for this whole module. And at some point we could fix it?
> > For user it would matter more that playbook just can be run again if it fails (not real idempotency meaning it only makes changes when it's not yet done). Is this now re-runnable?
> 
> While I agree with your remarks please note that this patch is in no way targeted at fixing idempotency in the scope of whole application.yml and wasn't supposed to, it's different topic. Idempotency matter here popped out only in the scope of the changed task code (helm charts archival)
> 
> Please raise a ticket for it if you'd like to see it fixed at some point.

Yes I'm completely aware of the topic of this commit, not so senile yet.
It's just our duty to comment these commits and this change decrease idempotence further more so why not to comment on that and not to decrease it.

This problem of application.yml is known from the beginning and we surely have a ticket from that. But wait...I just checked and did not found it????
So created it https://jira.onap.org/browse/OOM-2237

Patch-set: 7