Update patch set 7
Patch Set 7:

I'm not sure how much we want to test idempotency there but I've tested running that tar command twice and it rewrites existing tar file with newer data. It is same sized tar file but newer. If this is only thing what is important (to be able to run this command twice and not to create another tar file or add content to existing tar file) then its OK.

I'm rather have kibana pod fixed (even as temp solution, because archive module is already fixed but still not in stable release of ansible) then exclude it from HC and whitelist it as we want to keep kibana in onap-me for now.

Patch-set: 7
Reviewer: Gerrit User 5626 <5626@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=0