Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> > This needs to be tested for regressions on rhel platform (package
>  > build + deploy), onap-cge pipeline can be used for that.
> 
> is nighly stopped?
> 
> also we would need pipeline parameters for Ubuntu if not exists.

Nightly is stopped but it doesn't help as it tests master, not particular review and testing this review is what I meant.

As for job parameters I think that should be the last part of the exercise, for now it should be kept stable (job config) and all job tasks kept aside.

Patch-set: 4