Update patch set 1
Patch Set 1: Code-Review+1

This is funny: I've updated INFO.yaml but can't +2.
So, I'll be able to merge when I've got the rights

Again, I believe this is a bad way to do but acknowledge that it seems we are the only one thinking that.

Patch-set: 1
Reviewer: Gerrit User 2319 <2319@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=+1