Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> OK, that's what I thought...
> So Michal Ptacek, Samuli Silvius, Petr Ospaly, Tomas Levora, Milan Verespej, Bartlomiej Grzybowski, Michal Zegan are not working anymore on Offline Installer?
> 
> Maybe we should see how to change the committers list with actual committers also.

Just looked in the mirror and Bartlomiej Grzybowski is still me so can be ruled out from that list :-D
But as for the rest you are correct, they are no longer in the team.

Would it suffice if I just remove them from INFO.yaml or we need some nasty paperwork here with dozens of stamps?

Patch-set: 1