Update patch set 1
Patch Set 1:

> Patch Set 1:
> 

> So far I was under impression that to get commiters rights in a repo a candidate is obliged to have shown a commitment to that repo and with Konrad, Lasse and Krystian as only candidates right now we don't have such personalities. Do you think it would be possible to request commiters right for one of those mentioned just for the sake of having one spare commiter in our team? 
> 
> Krzysztof Opasiak is from Samsung as well but I just add him as reviewer for same reasons as I add You.

Yes, in order to have commiters right, you need to have proposed some commits and done some reviews.
So if you want that Konrad, Lasse or Krystian to become commiters, they'll need to do that for a month or two before asking for a vote.

As we're only four to be able to vote (the other active commiters), I'm pretty sure they would have it.
I'd also like to have more "diversity" in company but it's also hard to achieve...

Patch-set: 1